test: randomness 2 reshuflled
